La til radiogruppe i vurderingAarsavregningInngang.tsx for å fastsette at behandlingen har delt grunnlag
